updating after delete

+++ app/controllers/v1/clients_controller.rb
@@ -59,15 +59,21 @@ module V1
     # DELETE /clients/1
     # DELETE /clients/1.json
     def destroy
-      if current_user.try(:admin?)
-        @user = User.find(@client.user_id)
-      else
-        @user = current_user
-      end
+      @client.client_status = "archived"
 
-      @user.clients.delete_all
-      if @user.destroy
-        head(:ok)
+      if @client.save
+        if current_user.try(:admin?)
+          @user = User.find(@client.user_id)
+        else
+          @user = current_user
+        end
+
+        @user.clients.delete_all
+        if @user.destroy
+          head(:ok)
+        else
+          head(:unprocessable_entity)
+        end
       else
         head(:unprocessable_entity)
       end
